实现信息的发送与获取

+++ app/src/main/java/com/example/myapp/post/ui/message/chat/util/MsgBoxAdapter.java
@@ -71,8 +71,8 @@ public class MsgBoxAdapter extends RecyclerView.Adapter<MsgBoxAdapter.ViewHolder
                 intent.putExtra("cUid",chatUser.getId());
                 //传递信息记录
                 intent.putExtra("msgList", (ArrayList<Msg>) msgList);
-                //传递 msgboxID
-                intent.putExtra("MsgBoxId",msgBox.getMsgBoxId());
+                //传递 msgBoxID
+                intent.putExtra("MsgBoxId",msgBox.getMsgBoxId().intValue());
 
                 mContext.startActivity(intent);
